Updated openstack/openstack Project: openstack/nova  faeede29795d1eeee241d50f8521431e30ed540f
Add requested_destination field to RequestSpec

As agreed in the spec, we want to provide to the scheduler the possible
destination asked by the admin when calling evacuate or live-migrate.
Adding a possibility to default the field to None if an old RequestSpec
or a legacy dict wants to hydrate.

Change-Id: I6ddcaaca37fc5387c2d2e9f51c67ea9e85acb5c5
Partially-Implements: blueprint check-destination-on-migrations-newton