Updated openstack/openstack Project: openstack/nova  cfd64c976b8aaacaf09fd2a6303877674fcfada0
Verify requested_destination in the scheduler

Now that the RequestSpec provides a field for a requested destination, we
can know if the admin user provided a destination to verify (and not force
it). In order to verify this destination, we want to run the filters for
only this host, so that if the filters are not okay for that possibility,
the scheduler should raise a NoValidHost exception.

For the moment, nothing is setting this field but the next patch will provide
an API-level modification for putting it if the provided destination exists
(and changing the behaviour if the admin provides the host field for evacuate
and live-migrate)

Change-Id: Ifb803c708e89b7101ded48571e1af49c33117758
Partially-Implements: blueprint check-destination-on-migrations