Updated openstack/openstack Project: openstack/tempest  55fd59831c58998fe085cfc65e92872248cc4d26
Add hacking rule to enfore no config in tempest.lib

This commit adds a hacking rule to enforce that we never add a config
dependency on tempest/lib. Right now we're completely dependent on
reviewers catching this, it is a strong rule so we should ensure we
can't ever land a change that does this.

Change-Id: I1ab1ba52573c6706a50abcd021759c93dd19aa44