Updated openstack/openstack Project: openstack-infra/project-config  7f8816ef0efc92ae0298c358ed6b3ad498fb8131
Be more specific in Puppet OpenStack Zuul layout

Be more specific in Zuul layout rules for Puppet OpenStack jobs.
Our coverage matrix is quite complex because we run multiple scenarios
on multiple distros on multiple versions of Puppet on multiple versions
of OpenStack.

This patch makes the rules more robust so it won't match multiple cases
in the same time. Also add some documentation so we understand what we
expect for each rule.

Change-Id: Ie37a50fb4ae56a6d97956ac855565b1f68f12d77