Updated openstack/openstack Project: openstack/nova-specs  ca4facb3ec29f6bc2edbe5826467b64682a80881
Improve Scheduler Logging

The nova scheduler includes a number of very complicated filters with
non-obvious failure modes (the NUMATopologyFilter in particular comes
to mind). It is possible to have a situation where a given instances
fails to schedule, and it is not immediately apparent as to what
exactly caused the failure.

Accordingly, it is proposed that we allow for optional detailed
messages about precisely *why* a scheduler filter is failing.

Change-Id: I1789865cb9f16ca6f6e2416957529274c29d98fd