Updated openstack/openstack Project: openstack/nova-specs  ed3fa6328278e456b197981055f135d7b4186c06
Make newton-template.rst not a symlink

There doesn't seem to be any value to having a symlink to
mitaka-template.rst here. This makes newton-template.rst in specs/ it's
own file and updates the references to Mitaka. The symlinks in
newton/approved and newton/implemented are left in place as we don't
need to duplicate the file in the repo.

Change-Id: I22837d72b8d9d525511dad1e170a923aea094a8b