Updated openstack/openstack Project: openstack/ironic  3559f151520fd88a91c3ad6f165a483b0f383884
NOTIFICATION_TRANSPORT should be global

This is a followup to "Use messaging notifications transport
instead of default" (Ic72fd658fe9b35d39cc3d7da2e4b9fe3454edfba).

The use of NOTIFICATION_TRANSPORT should be global. A unit
test is added; it fails if the variable isn't global.

A typo was addressed in the release notes.

Change-Id: Ie303d79aba893b36a7aecb29c0ff2703f742146e