Updated openstack/openstack Project: openstack/oslo.config  bb0f7e3880bd83dbb9088f3f8045152ce8e7897b
generator: format string default value for List type properly

List type allows to pass string values as default values, which are then
handled by splitting on commas. For formatting matters, if the default
value is of string type, there is no special conversion needed, and we
can reuse the value as-is.

Change-Id: Ie9ecdfc064e886f6138b1fa9e839d8f95a7df339
Closes-Bug: #1586366