Updated openstack/openstack Project: openstack/nova-specs  dd167ad2c9b1b27aaaf46918873161c89453a57c
Libvirt: Use the virtlogd deamon

If the *serial console* feature is enabled on a compute node with
``[serial_console].enabled = True`` it deactivates the logging of the
boot messages. From a REST API perspective, this means that the two
APIs ``os-getConsoleOutput`` and ``os-getSerialConsole`` are mutually
exclusive. Both APIs can be valuable for cloud operators in the case
when something goes wrong during the launch of an instance. This
blueprint wants to lift the XOR relationship between those two REST
APIs.

blueprint libvirt-virtlogd

Change-Id: I9a1fbf005b0f48df90093a346579d9ddc64f7846