Updated openstack/openstack Project: openstack/sahara  46df5c685dd4b6c98aebd81eb2c6e5aad5d1ce89
Readable logging for Heat templates

Sahara used to write Heat template to log output in format
which cannot be parsed with JSON or YAML validation tools.
As a result it caused difficultis for a user to read the
initial template.

This issue is solved by making  Sahara write logs in json format.

Closes-bug: 1552239

Change-Id: Ie7a9f6fcad29899dd53fa43884e783c58ea304ec