Updated openstack/openstack Project: openstack/requirements  e00676a8b2b2292138f3f02c2b3b949573730a49
Raise fixtures requirement to 3.0.0

Fixtures 1.4 has a bug with MonkeyPatch and its handling of
staticmethods for python3. Essentially a monkey patched static method is
no longer a static method after the fixture cleans up.

This was fixed in fixtures 2.0 but unfortunately that release
inadvertently added further constraints on the use of MonkeyPatch that
caused a number of legitimate uses to error. So fixtures was capped to
1.4.

Fixtures 3.0 fixes the bug from 1.4 but does not cause failures with
current legitimate uses of MonkeyPatch. This has been confirmed against
py27 and py24 unit tests on Nova and Neutron, and functional tests on
Nova.

Change-Id: I214583f7c27fcfe3f6b17700710ef27c8ecdbbab
Depends-On: I9c5635d0930b091e1ae0a1439d687dd20a0ff939