Updated openstack/openstack Project: openstack-infra/project-config  068cfa5a53d44552943332ed4ce9197277fa9027
Fix translation proposals

Translation proposals have been changing files in the VENV directory
since the pull_from_zanata function downloads all files and it's not
clear which directories to use. The find there is from the top level.

Move the VENV to outside of the git tree to the tmp directory - like
it's done in all other cases for jenkins/scripts/. This also allows to
simplify the upstream_translation_update script.

Change-Id: Idf344d89691189c7c9ed46ef885433d3db14dff2