Updated openstack/openstack Project: openstack/openstackdocstheme  157770246dc7bda2a1b92f77d106c1a8b2ab20e2
Fix text font-family of admonitions.

This prepends a div with the correct font-awesome classes,
and leaves the title (Info / Warning etc) to inhert
fonts from the rest of the document.

This also removes the need for the prepended space to the title.
(as the content of the div is a space).

Change-Id: I39b9c666534ccec4a0c801c30d01cf10e68bfba0