Updated openstack/openstack Project: openstack/nova  850d8cdc2cb8163285bdacd1ea83b4fe145e8d16
remove /v2.1/{tenant_id} from all urls

As discussed at summit, the version part of the URL is not really
relevant, or a thing a user should be filling out themselves, this
should instead be set by the service catalog and extracted from the
token.

This removes it's reference in all documented REST urls, and adds a
new section describing how one gets the base URL for all calls.

Change-Id: I4306b8c3de0225e54f3909dd8a1fb293c4e5944c