Updated openstack/openstack Project: openstack/openstackdocstheme  ac0cde638814b39cabff224f309c495a93951fb2
Set side bar content to be configurable

Sometimes (e.g. for the api-ref pages) the TOC should be in the side bar.
This allows this to be set as a theme option, but defaulting to
current behaviour - the new TOC will only be set
if a project sets the "sidebar_mode" option to "toc".

Change-Id: I9b93b671d4c32f5b237b0d79ae2be39059402d72