Updated openstack/openstack Project: openstack/keystone  de516b9552b464b90e11d2298ecf2630264af48f
Cache service providers on token validation

On fernet token validation we rebuild the entire token catalog. This includes
rebuilding the service providers references in the token if using a federated
token. In the mitaka release, patches were added to cache user catalogs and
role assignments. This allows us to do the same thing for service providers,
which decreases the time spent in get_token_data() by 34%.

Co-Authored-By: Dolph Mathews <dolph.mathews@gmail.com>

Related-Bug: 1489061
Change-Id: I3dfe66880295ef36a18ca3e384191106b74be795