Updated openstack/openstack Project: openstack/horizon  cf541e899875c5e4ddb40cbf045efe0d41ef0fd0
Use JQuery promise/deferred objects to track AJAX

The JQuery library offers objects known as deferred for tracking
multiple call backs, this can be used with AJAX to track cases when we
only want to do something once multiple requests are complete such as
re-calling the datatables update method.

This also allows for us to track the completed state of a request in the
ajax queue without injecting a custom completed callback into every
request that is handled by the queue.

Change-Id: I7d5acb6d05a5822b2de7d73678797166276ac599
Closes-Bug: 1442146