Updated openstack/openstack Project: openstack/openstack-manuals  7b88b45272f68dd2b6f14051c2da3843d9b2d0f2
Revert "[config-ref] Remove cisco FC Zone manager driver"

The Cisco FCZM driver was removed for lack of CI during the Mitaka
development cycle. Prior to the final Mitaka release this was
rectified and the FCZM driver was added back in to the release.
This cisco FCZM driver doc should be added again

This reverts commit c2ed9b27819043c0b7e109bbb355cf3b36ed7c05.

backport: mitaka

Change-Id: Ibe0d72d1070fcc49194c488806c71a0423df790b