Updated openstack/openstack Project: openstack/python-openstackclient  55a7ba890d9c3a53b7728efe41053b79ba1e60d3
Fix --enable options on commands

The --enable option on commands is ignored when the arguments are parsed.
This is related to the --enable-beta-commands option. Renaming the option
to --os-beta-command fixes the problem.

There's no need to handle backwards compatibility for the option name
change because there hasn't been an OSC release yet with beta commands.

Change-Id: I0327ba8a2058858a83e9a42e231470ed733cc834
Closes-Bug: #1588384