Updated openstack/openstack Project: openstack/python-openstackclient  81718beb8f2fa891b10c0a18e3378188daed9f7d
Fix wrong test in flavor unit tests

In test_flavor_set_no_project(), we aimed to test a
situation like this: User specifies "--project" option,
but didn't specifies the project name or ID.

But in the source code, it becomes "--project ''".

The test could past because if project and property
are both None, the command will raise an exception.

Change-Id: I39567306debb901e8bad420fa2492f1b207efddc