Updated openstack/openstack Project: openstack/cinder  3f9800afcfc4eb03a7045267d31c4d2655b0210c
Storwize: terminate_connetion NPIV perf improvement

When terminate connection is called on a VM that is using
NPIV connections while the VM is shutdown such as when the
the VM is being deleted, the NPIV ports are no longer logged
into the fabric and the lsfabric fast path will not find the VM.
Add a fast path this NPIV case is to first check the host names
from the list of hosts that have mapped the volume before
exhaustively searching the entire host list for a match

Fixed an issue with the simulator not resetting lshost after
the simulated condition fired.

Added a new UT to cover the code change.

Change-Id: Ie38fb708ad7eae62261e71ab0bcb505c3efdb3b1
Closes-Bug: #1580342