Updated openstack/openstack Project: openstack/ceilometer  0569b75b107d1c03aa8e06bbe566aedcf770d7cf
Drop timestamping in pollsters

Many of the pollsters define the timestamp individually for each
generated sample, which should really be timestamping based on when
the data was polled and not when each sample is generated. We need to
set the timestamp of the polled data to the timestamp when the polling
starts for unity.

Previous commit be3f7fb8300e8853ebd35986cdf159422aa247a already unified
the timestamp in samples, and now drop the timestamping in pollsters.

Change-Id: I9c3047e70fcd6c02278dee91603e2aeb2ca8b8c8
Closes-Bug: #1491509
Implements: blueprint unify-timestamp-of-polled-data