Updated openstack/openstack Project: openstack/keystonemiddleware  a4d4eb813f505f385aa7c70b597baf8ccfe3aa86
Improve documentation for auth_uri

The auth_uri configuration option causes a lot of confusion next to the
auth_url option in auth plugins. In the short term, the auth_uri option
deserves some better documentation, and in the long term, I included an
inline comment describing an approach to eliminate the option
altogether.

Change-Id: Ied0ba0abc906b06e36cf60ba3d4bfc0ba9a6fdac