Updated openstack/openstack Project: openstack/python-heatclient  8ff8e949525ea2a4f0094c3038421911f9f1f4d9
A resource list formatter for graphviz dot diagrams

This change registers a "dot" formatter for the stack resource list
command. It generates the text for a graphviz dot diagram which can be
rendered as an image with this example usage:

  openstack stack resource list --format dot -n3 my_stack | dot -Tsvg -o my_stack.svg

Nested resources are fully supported by creating nodes for top-level
resources then linking stack resources to a subgraph representing the
nested stack.

Change-Id: I4b899287eaf818137d60cb278db2d86598aa6794