Updated openstack/openstack Project: openstack/nova  2bd6b6be8b8e42a1c9bc39de627f817f86d11743
api-ref method verification for os-cells

Confirm and correct list methods and response codes in os-cells.

This is mostly just checking, correcting and adding methods, but
in the process of doing that it became clear that there was quite
a bit missing so several TODO notes have been left in place so that
the next stages of work are a bit more obvious.

* 405 has been removed everywhere as we don't send that response
* 400 badRequest is required anywhere, as that's what you'll get if
  tenant_id does not match the context's project_id
* 503 does not happen as there's no application controlled proxying
  in this api
* the URL for listing cells with details was incorrect
* the example response for cells with details was an empty list,
  this is now a todo suggesting an as yet non-existent file
* capacities was not listed, now is
* not clear what capacities really are, so a TODO
* creating a new cell (via POST) was not there, so that is now added
  as a method
* for that POST, request and response bodies, parameters currently a
  TODO
* Update (PUT) and DELETE of a cell has been added, but details are TODO

Part of bp:api-ref-in-rst

Change-Id: I3f123f821820e5a5a4deff61fb716d43dc486142