Updated openstack/openstack Project: openstack/nova  d08cff787507d546cc400bd1d2f3c910580fd86c
[libvirt] Incorrect parameters passed to migrateToURI3

migrateToURI3 does not take 'bandwidth' argument so we shouldn't try
to pass it. Actually it results with exception when libvirt>=1.2.17
is used.

Note the bandwidth parameter is already in the params list via the call
from the libvirt driver in Nova so we don't have to check to see if it's
already in 'params'. However, this is ugly and leaks libvirt API info
out of the Guest object. That will be refactored and cleaned up in a
follow-on change.

Closes-Bug: #1589548

Change-Id: Ieea2d955c763655e2b2abfb99fccb1a4c7169794