Updated openstack/openstack Project: openstack/nova  73f0e3058ced1f4b30a465b65332829d85dc0ed9
Remove code referencing inventory table in cell DB

The ComputeNode.save() and create() methods were still looking for
fields that began with the prefix inv_ and treating those differently.
Those fields are no longer in the return from compute_node_get() since
the last patch reverted the code that joins compute_nodes to allocations
and inventories within the cell DB.

This patch undoes the calls to compute_node_get() that were done to grab
that inventory information which should reduce the number of DB calls
when compute nodes are saved or created.

Change-Id: I5856c9e52c74e4f4530846f76f0a9d2735af23dc
Closes-bug: #1572555