Updated openstack/openstack Project: openstack-infra/system-config  3a694b3d2eee1ba32baac8ad1c0deebbc08945c2
Set openstack-health cache_expiration to 5min

This commit shrinks the openstack-health cache expiration time from
the default 15min to 5min. Since we added the async cache updating
the need for a longer cache period has gone away. We had a conservative
default originally to make it less likely that a user would hit a cache
refresh, which with ES queries could take a long time. The async worker
always makes sure that the cached result is always used and if it's
stale the cache is refreshed in the background so there is no user
facing delays. (after the initial cache population)

Change-Id: Ice6ca0e61504f3fff5c4d1e863811710c7c89fb7