Updated openstack/openstack Project: openstack/cinder  82c352649bb86b6854bfbf70f94dca45bbbc8b3e
Dell SC: Added support for failover_host failback

The Dell SC driver did not support failover_host failback. If
failover_host is run with a target of "default" the Dell SC
driver will attempt to fail back to the original SC. If unable
to failback a volume the driver will mark the volume in error
but so long as the SC is available will eventually return control
to the original SC.

At conclusion of the failback the original volume should be
replicated to all configured replication targets and contain
the latest changes from the backend the volume was failed over
to.

This can take quite some time as replications need to be brought
up, synced, torn down and re-created in the opposite direction.

Added logging line to retype. Fixed a retype option.

Some find_volume calls fixed.

Change-Id: I5c12119ca9604ae1c4b167743f12621d2cd99075