Updated openstack/openstack Project: openstack/diskimage-builder  13323c9495a932ff4c84d939a2457ffbc300a97f
Export die() function

I realised I'd been using die() in a few places assuming it was
available, but it wasn't exported.  I guess it didn't matter because
whatever was wrong, we were failing anyway :)

This exports the function to make it available to sub-processes, which
should remove the need to source it as done in several places.

Change-Id: I7b9a5a6db406e160099b6ed9fde80455ae227327