Updated openstack/openstack Project: openstack/nova  3b1054d339d505a47d55ce1b1c7dae3eda0dc8a1
Don't chown a config disk which already exists

Change I20bafc3482ead30297ad5402a09e97141816f4d1 fixed a bug when
trying to recreate a config disk which already exists. If the disk
already exists and its guest has previously booted, libvirt will have
chowned it. This causes a permission denied error when trying to
overwrite it.

We no longer overwrite config disks, so this is no longer necessary.
It is also inconsistent with our handling of other disks, so we remove
it.

Change-Id: Idcaf8e5dbd2dbf6373cb5d79e0822fbff3664703