Updated openstack/openstack Project: openstack/ceilometer  3b42b2514f467c0301f48ccfa82c2859a184fedf
Copy images_client from tempest

Now the gate test of ceilometer is broken, because the ceilometer test
is using nonstable tempest interfaces and these interfaces are changed.
QA team defines interfaces which are not under tempest.lib as unstable
clearly according to

https://github.com/openstack/tempest/blob/master/doc/source/library.rst#stability

So current test way is against the QA policy, so here is a workaround.
Copy&paste is not good, almost wrong. However, if still continuing the
tests, it would be a nice option to copy&paste and pass the gate test.
After QA team defines stable interfaces, that is a nice time to switch
using the stable interfaces with removing this copy&paste code.

Closes-Bug: #1589426

Depends-On: I3abd9049560ee507b3610ab482c697a239f13a3b
Change-Id: I30266f1b690ea105511d9b1162755bcc85e2e69a